requirements.txt
absl-py==0.11.0
albumentations==0.5.2
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.14.2
grpcio==1.34.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
np-utils==0.5.12.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
pipreqs==0.4.10
playsound==1.2.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyobjc==7.0.1
pyobjc-core==7.0.1
pyobjc-framework-Accounts==7.0.1
pyobjc-framework-AddressBook==7.0.1
pyobjc-framework-AdSupport==7.0.1
pyobjc-framework-AppleScriptKit==7.0.1
pyobjc-framework-AppleScriptObjC==7.0.1
pyobjc-framework-ApplicationServices==7.0.1
pyobjc-framework-AuthenticationServices==7.0.1
pyobjc-framework-AutomaticAssessmentConfiguration==7.0.1
pyobjc-framework-Automator==7.0.1
pyobjc-framework-AVFoundation==7.0.1
pyobjc-framework-AVKit==7.0.1
pyobjc-framework-BusinessChat==7.0.1
pyobjc-framework-CalendarStore==7.0.1
pyobjc-framework-CFNetwork==7.0.1
pyobjc-framework-CloudKit==7.0.1
pyobjc-framework-Cocoa==7.0.1
pyobjc-framework-Collaboration==7.0.1
pyobjc-framework-ColorSync==7.0.1
pyobjc-framework-Contacts==7.0.1
pyobjc-framework-ContactsUI==7.0.1
pyobjc-framework-CoreAudio==7.0.1
pyobjc-framework-CoreAudioKit==7.0.1
pyobjc-framework-CoreBluetooth==7.0.1
pyobjc-framework-CoreData==7.0.1
pyobjc-framework-CoreHaptics==7.0.1
pyobjc-framework-CoreLocation==7.0.1
pyobjc-framework-CoreMedia==7.0.1
pyobjc-framework-CoreMediaIO==7.0.1
pyobjc-framework-CoreMIDI==7.0.1
pyobjc-framework-CoreML==7.0.1
pyobjc-framework-CoreMotion==7.0.1
pyobjc-framework-CoreServices==7.0.1
pyobjc-framework-CoreSpotlight==7.0.1
pyobjc-framework-CoreText==7.0.1
pyobjc-framework-CoreWLAN==7.0.1
pyobjc-framework-CryptoTokenKit==7.0.1
pyobjc-framework-DeviceCheck==7.0.1
pyobjc-framework-DictionaryServices==7.0.1
pyobjc-framework-DiscRecording==7.0.1
pyobjc-framework-DiscRecordingUI==7.0.1
pyobjc-framework-DiskArbitration==7.0.1
pyobjc-framework-DVDPlayback==7.0.1
pyobjc-framework-EventKit==7.0.1
pyobjc-framework-ExceptionHandling==7.0.1
pyobjc-framework-ExecutionPolicy==7.0.1
pyobjc-framework-ExternalAccessory==7.0.1
pyobjc-framework-FileProvider==7.0.1
pyobjc-framework-FileProviderUI==7.0.1
pyobjc-framework-FinderSync==7.0.1
pyobjc-framework-FSEvents==7.0.1
pyobjc-framework-GameCenter==7.0.1
pyobjc-framework-GameController==7.0.1
pyobjc-framework-GameKit==7.0.1
pyobjc-framework-GameplayKit==7.0.1
pyobjc-framework-ImageCaptureCore==7.0.1
pyobjc-framework-IMServicePlugIn==7.0.1
pyobjc-framework-InputMethodKit==7.0.1
pyobjc-framework-InstallerPlugins==7.0.1
pyobjc-framework-InstantMessage==7.0.1
pyobjc-framework-Intents==7.0.1
pyobjc-framework-IOSurface==7.0.1
pyobjc-framework-iTunesLibrary==7.0.1
pyobjc-framework-LatentSemanticMapping==7.0.1
pyobjc-framework-LaunchServices==7.0.1
pyobjc-framework-libdispatch==7.0.1
pyobjc-framework-LinkPresentation==7.0.1
pyobjc-framework-LocalAuthentication==7.0.1
pyobjc-framework-MapKit==7.0.1
pyobjc-framework-MediaAccessibility==7.0.1
pyobjc-framework-MediaLibrary==7.0.1
pyobjc-framework-MediaPlayer==7.0.1
pyobjc-framework-MediaToolbox==7.0.1
pyobjc-framework-Metal==7.0.1
pyobjc-framework-MetalKit==7.0.1
pyobjc-framework-MetalPerformanceShaders==7.0.1
pyobjc-framework-ModelIO==7.0.1
pyobjc-framework-MultipeerConnectivity==7.0.1
pyobjc-framework-NaturalLanguage==7.0.1
pyobjc-framework-NetFS==7.0.1
pyobjc-framework-Network==7.0.1
pyobjc-framework-NetworkExtension==7.0.1
pyobjc-framework-NotificationCenter==7.0.1
pyobjc-framework-OpenDirectory==7.0.1
pyobjc-framework-OSAKit==7.0.1
pyobjc-framework-OSLog==7.0.1
pyobjc-framework-PencilKit==7.0.1
pyobjc-framework-Photos==7.0.1
pyobjc-framework-PhotosUI==7.0.1
pyobjc-framework-PreferencePanes==7.0.1
pyobjc-framework-PushKit==7.0.1
pyobjc-framework-Quartz==7.0.1
pyobjc-framework-QuickLookThumbnailing==7.0.1
pyobjc-framework-SafariServices==7.0.1
pyobjc-framework-SceneKit==7.0.1
pyobjc-framework-ScreenSaver==7.0.1
pyobjc-framework-ScriptingBridge==7.0.1
pyobjc-framework-SearchKit==7.0.1
pyobjc-framework-Security==7.0.1
pyobjc-framework-SecurityFoundation==7.0.1
pyobjc-framework-SecurityInterface==7.0.1
pyobjc-framework-ServiceManagement==7.0.1
pyobjc-framework-Social==7.0.1
pyobjc-framework-SoundAnalysis==7.0.1
pyobjc-framework-Speech==7.0.1
pyobjc-framework-SpriteKit==7.0.1
pyobjc-framework-StoreKit==7.0.1
pyobjc-framework-SyncServices==7.0.1
pyobjc-framework-SystemConfiguration==7.0.1
pyobjc-framework-SystemExtensions==7.0.1
pyobjc-framework-UserNotifications==7.0.1
pyobjc-framework-VideoSubscriberAccount==7.0.1
pyobjc-framework-VideoToolbox==7.0.1
pyobjc-framework-Vision==7.0.1
pyobjc-framework-WebKit==7.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.12.8
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
